Split out filters
