+++ lib/kramdown-latex.rb
@@ -1,56 +1,4 @@
-require 'set'
 require 'kramdown/converter'
-
-Class.new(Nanoc::Filter) do
-  identifier :latex2pdf
-  type :text => :binary
-
-  def run(content, params = {})
-    unless system('which', 'xelatex', out: '/dev/null')
-      $stderr.puts "Warning: `xelatex` not found; PDF generation disabled."
-      File.write(output_filename, '')
-      return
-    end
-
-    Tempfile.open(['nanoc-latex', '.tex']) do |f|
-      f.write(content)
-      f.flush
-
-      3.times do
-        system('xelatex', '-halt-on-error', '-output-directory', File.dirname(f.path), f.path)
-      end
-
-      system('mv', f.path.sub('.tex', '.pdf'), output_filename)
-    end
-  end
-end
-
-Class.new(Nanoc::Filter) do
-  identifier :kramdown_latex
-
-  def run(content, params = {})
-    params = params.merge(
-      latex_headers: %w(chapter section subsection subsubsection paragraph subparagraph),
-      input: @item.identifier.ext == 'erb' ? 'html' : 'kramdown',
-    )
-
-    document = ::Kramdown::Document.new(content, params)
-
-    document.warnings.each do |warning|
-      $stderr.puts "kramdown warning: #{warning}"
-    end
-
-    document.to_nanoc_ws_latex
-  end
-end
-
-Class.new(Nanoc::Filter) do
-  identifier :remove_pre_code_language
-
-  def run(content, params = {})
-    content.lines.reject { |l| l =~ /^\s+#!/ }.join
-  end
-end
 
 module Kramdown
   module Converter
